// src/app/administration/administration.interface.ts
export interface Administration {
    no: string;
    date: string;
    area: string;
    locality: string;
    detail: string;
    description: string;
    latitude: string;
    longitude: string;
    observations: string;
    recommendationsNavigator: string;
    source: string;
    preparedBy: string;
}